requirements.txt
nltk
streamlit
tenacity
openai
praw
transformers
python-dotenv